+++ app-api/src/main/java/org/yechan/service/user/SellerPersister.java
@@ -1,4 +1,4 @@
-package org.yechan.service;
+package org.yechan.service.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
